LPS-157265 UpgradeModulesFactory
