+++ portal-impl/src/com/liferay/portal/upgrade/util/UpgradeModulesFactoryImpl.java
@@ -1,0 +1,54 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.upgrade.util;
+
+import com.liferay.portal.kernel.upgrade.util.UpgradeModules;
+import com.liferay.portal.kernel.upgrade.util.UpgradeModulesFactory;
+
+/**
+ * @author Arthur Chan
+ */
+public class UpgradeModulesFactoryImpl implements UpgradeModulesFactory {
+
+	@Override
+	public UpgradeModules create(
+		String[] bundleSymbolicNames, String[][] convertedLegacyModules) {
+
+		return new UpgradeModules() {
+
+			@Override
+			public String[] getBundleSymbolicNames() {
+				if (bundleSymbolicNames == null) {
+					return new String[0];
+				}
+
+				return bundleSymbolicNames;
+			}
+
+			@Override
+			public String[][] getConvertedLegacyModules() {
+				if ((convertedLegacyModules == null) ||
+					(convertedLegacyModules[0] == null)) {
+
+					return new String[0][0];
+				}
+
+				return convertedLegacyModules;
+			}
+
+		};
+	}
+
+}
