+++ portal-kernel/src/com/liferay/portal/kernel/upgrade/util/UpgradeModulesFactoryUtil.java
@@ -1,0 +1,41 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.kernel.upgrade.util;
+
+/**
+ * @author Arthur Chan
+ */
+public class UpgradeModulesFactoryUtil {
+
+	public static UpgradeModules create(
+		String[] bundleSymbolicNames, String[][] convertedLegacyModules) {
+
+		return _upgradeModulesFactory.create(
+			bundleSymbolicNames, convertedLegacyModules);
+	}
+
+	public static UpgradeModulesFactory getUpgradeModulesFactory() {
+		return _upgradeModulesFactory;
+	}
+
+	public void setUpgradeModulesFactory(
+		UpgradeModulesFactory upgradeModulesFactory) {
+
+		_upgradeModulesFactory = upgradeModulesFactory;
+	}
+
+	private static UpgradeModulesFactory _upgradeModulesFactory;
+
+}
